move extract cmd tests to own module
